+++ src/opnieuw/src/handler/pipeline/human_engine/scores/requests.rs
@@ -6,7 +6,7 @@ use crate::GA;
 
 impl RequestContext {
     pub fn check_traffic_ip(&self) -> u32 {
-        let mult = (self.ip.points.get() as f64 / 5000 as f64) as u32 * 100;
+        let mult = (self.ip.points.get() as f64 / 5000_f64) as u32 * 100;
 
         if mult > 50 {
             GA.handler.he.request_alert.inc();
